worker: refactor out back refs
